refactor: move LoadingProgressBar to components package
