+++ identity_presentation/src/commonMain/kotlin/net/thechance/mena/identity/presentation/screen/addresses/myAddresses/AddressesScreen.kt
@@ -26,7 +26,7 @@ import net.thechance.mena.designsystem.presentation.theme.theme.Theme
 import net.thechance.mena.identity.presentation.base.BaseScreen
 import net.thechance.mena.identity.presentation.components.AddressSnackBar
 import net.thechance.mena.identity.presentation.components.NoSavedLocationsLayout
-import net.thechance.mena.identity.presentation.components.util.LoadingProgressBar
+import net.thechance.mena.identity.presentation.components.LoadingProgressBar
 import net.thechance.mena.identity.presentation.screen.addresses.addEditLocation.AddEditLocationScreen
 import net.thechance.mena.identity.presentation.screen.addresses.myAddresses.components.AddressCard
 import net.thechance.mena.identity.presentation.screen.addresses.myAddresses.components.MyAddressesAppBar
